Fix pulling opening geometry.
Make toolkit compliant by moving Adapter only methods to the Adapter
