+++ DIALux_Adapter/CRUD/Create.cs
@@ -28,9 +28,6 @@ using System.Threading.Tasks;
 
 using BH.oM.Base;
 
-using BH.oM.Adapters.DIALux;
-using BH.Engine.Adapters.DIALux;
-
 using BH.oM.Environment.Elements;
 using BH.Engine.Environment;
 
@@ -65,7 +62,7 @@ namespace BH.Adapter.DIALux
             return true;
         }
 
-        private void WriteVersion(StreamWriter sw, BH.oM.Adapters.DIALux.Version version)
+        private void WriteVersion(StreamWriter sw, Version version)
         {
             sw.WriteLine("[VERSION]");
             sw.WriteLine("STFF=" + version.VersionNo);
