+++ DIALux_Adapter/CRUD/Read.cs
@@ -31,7 +31,6 @@ using BH.Engine.Adapter;
 using BH.oM.Adapter;
 using BH.oM.Base;
 using System.IO;
-using BH.Engine.Adapters.DIALux;
 
 namespace BH.Adapter.DIALux
 {
