+++ DIALux_Adapter/Convert/Environment/OpeningType.cs
@@ -34,16 +34,13 @@ using BH.oM.Environment.Elements;
 using BH.Engine.Environment;
 using BH.Engine.Geometry;
 
-using BH.oM.Adapters.DIALux;
-
-namespace BH.Engine.Adapters.DIALux
+namespace BH.Adapter.DIALux
 {
     public static partial class Convert
     {
         [Description("Convert a BHoM Environment Opening Type into a DialUX Furnishing Type")]
         [Input("openingType", "A BHoM Environment Opening Type to convert")]
         [Output("furnishingType", "A DialUX furnishing type")]
-        [PreviousVersion("4.0", "BH.Engine.Adapters.DIALux.Convert.ToDialUX(BH.oM.Environment.Elements.OpeningType)")]
         public static string ToDIALux(this OpeningType openingType)
         {
             switch(openingType)
