+++ DIALux_Adapter/Objects/Structure.cs
@@ -29,17 +29,15 @@ using System.Threading.Tasks;
 using BH.oM.Base;
 using BH.oM.Geometry;
 
-namespace BH.oM.Adapters.DIALux
+namespace BH.Adapter.DIALux
 {
-    public class Room : BHoMObject
+    public class Structure : BHoMObject
     {
-        public virtual double Height { get; set; } = 0.0;
-        public virtual string Description { get; set; } = "BHoM Created Room";
-        public virtual List<Point> Points { get; set; } = new List<Point>();
-        public virtual List<Structure> Structures { get; set; } = new List<Structure>();
-        public virtual List<Luminaire> Luminaires { get; set; } = new List<Luminaire>();
-        public virtual List<Furnishing> Furnishings { get; set; } = new List<Furnishing>();
-        public virtual string SpecificConnectedLoad { get; set; } = "";
-        public virtual string MeanLuxWorkingPlane { get; set; } = "";
+        public virtual string Type { get; set; } = "";
+        public virtual Point Position { get; set; } = new Point();
+        public virtual double RotationX { get; set; } = 0.0;
+        public virtual double RotationY { get; set; } = 0.0;
+        public virtual double RotationZ { get; set; } = 0.0;
+        public virtual Luminaire LuminaireStructure { get; set; } = new Luminaire();
     }
 }
