--- DIALux_oM/Versioning_32.json
@@ -1,18 +1,0 @@
-{
-  "Namespace": {
-    "ToNew": {
-      "BH.Engine.XML":  "BH.oM.Adapters.DIALux"
-    },
-    "ToOld": {
-      "BH.Engine.Adapters.XML": "BH.oM.DIALux"
-    }
-  },
-  "Type": {
-    "ToNew": {
-      
-    },
-    "ToOld": {
-      
-    }
-  }
-}
